[feat/#33] DetailContent에선 5개 리뷰만 보이게 변경
